+++ src/Database/Model.php
@@ -2,10 +2,13 @@
 
 namespace Butler\Service\Database;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model as BaseModel;
 
 abstract class Model extends BaseModel
 {
+    use HasFactory;
+
     protected $connection = 'default';
 
     protected $guarded = [];