+++ src/Providers/RouteServiceProvider.php
@@ -8,55 +8,20 @@ use Illuminate\Support\Facades\Route;
 class RouteServiceProvider extends ServiceProvider
 {
     /**
-     * This namespace is applied to your controller routes.
-     *
-     * In addition, it is set as the URL generator's root namespace.
-     *
-     * @var string
-     */
-    protected $namespace = 'App\Http\Controllers';
-
-    /**
-     * Define the routes for the application.
+     * Define your route model bindings, pattern filters, etc.
      *
      * @return void
      */
-    public function map()
+    public function boot()
     {
-        $this->mapApiRoutes();
+        $this->routes(function () {
+            if (is_file($apiRoutes = base_path('routes/api.php'))) {
+                Route::middleware('api')->group($apiRoutes);
+            }
 
-        $this->mapWebRoutes();
-    }
-
-    /**
-     * Define the "web" routes for the application.
-     *
-     * These routes all receive session state, CSRF protection, etc.
-     *
-     * @return void
-     */
-    protected function mapWebRoutes()
-    {
-        if (is_file($routes = base_path('routes/web.php'))) {
-            Route::middleware('web')
-                 ->namespace($this->namespace)
-                 ->group($routes);
-        }
-    }
-
-    /**
-     * Define the "api" routes for the application.
-     *
-     * These routes are typically stateless.
-     *
-     * @return void
-     */
-    protected function mapApiRoutes()
-    {
-        if (is_file($routes = base_path('routes/api.php'))) {
-            Route::middleware('api')
-                ->namespace($this->namespace)
-                ->group($routes);
-        }
+            if (is_file($webRoutes = base_path('routes/web.php'))) {
+                Route::middleware('web')->group($webRoutes);
+            }
+        });
     }
 }