+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## Changed
+- Require Laravel 8.
+
 ## [0.2.13] - 2020-08-03
 
 ### Added